YouAreSpecial (Remove creation of directory with "special permissions".)
After cloning the EntityFramework repository and installing SQL Server 2012 three tests were failing. I traced this down to the creation of a directory with special permissions. After removing this code all the tests now pass. Andrew and I are not sure why this code was there but it may have been due to the way SQL Express was installed on the DevDiv VM. We don’t think it is needed, and even if it is, then the way it is right now seems broken since it doesn’t created a directory with full access than can then be used by the logged-in user. Initial fix is just to delete the code. If it turns out that something is needed here then we will have to revisit.