Port from EF5:
ItLooksWalksQuacksLikeAConnectionStringBut... (Add LocalDbConnectionFactory.)

Add a LocalDbConnectionFactory to configure LocalDb without needing to pass a connection string to the SqlConnectionFactory constructor. This has a couple of advantages:
•	We can use AttachDbFilename, which aligns with ASP.NET, without a breaking change. AttachDbFilename is only used if the app domain DataDirectory is set, otherwise it appears to not work well with LocalDb.
•	The config file now doesn’t have something in it that looks like a connection string and thereby confuses people.

The config file manipulation has been changed to use this new connection factory and also to call the parameterless constructor of the SqlConnectionFactory for SQL Express. This only happens when adding new entries; existing factory config is not changed.

Tests have been added for the new/changed code and I also tested the package install on a SQL Express machine and a LocalDb machine.