DonutsForTheFxCop (Making code run clean with FxCop.)
There are a lot of changes here. Most of them are done with ReSharper and should be safe, but I'm going to make a pass through them and I would appreciate if someone else does do.

I will file a bug for the suppressions of not calling Dispose since I'm not sure that we should be suppressing in all places but it will take some digging to figure it all out.

I plan to run ReSharper code cleanup but didn't want to include those changes in this checkin.

I have tried not to break the public surface of the assembly. I have not worried about breaking the internal surface--i.e. internal properties and methods. This could break assemblies compiled with InternalsVisibleTo, but technically we don't have any of those yet (except the tests) due to the assembly merge and hopefully will never have any.