LocalDbAndTheCity (Fix the broken CI.)
The problem is that when the test runs the first time it creates a LocalDb database with the MDF file in a temp location provided by Team City. This location is then cleaned, causing the MDF file to be deleted. The next time the test runs LocalDb still expects the MDF to be in the temp location. Since it isn’t, LocalDb throws.

The fix is to set these tests up so that they never leave LocalDb databases around after the test is completed. We don’t generally do this because it makes debugging harder, especially for occasional failures, if the database is always deleted after the test finishes and also it makes the tests slower because databases always have to be re-created. But in this case it seems like the best option.