WeGotFunkyFunctions (Make EntityFunctions work with DbContext and add tests.)
The code in EntityFunctions that bootstraps a LINQ query when the function is called directly as opposed to being in an expression tree did not recognize DbQuery and hence failed. This has been fixed so that both ObjectQuery and DbQuery are recognized. Tests have been added to provide coverage of all EntityFunctions used with both DbContext and ObjectContext. This will allow me to change the name/namespace of both EntityFunctions and EdmFunctionAttribute with confidence.

In addition the code in EntityFunctions has been cleaned up and refactored. I have added documentation based mainly on the documentation already in MSDN so that this is now inline with the code going forward.