FreeCup<CQT>ForThePublic (Make the remainder of the public CQT work...err...public!)
My understanding is that some of the public CQT work we did for .NET 4 was not made public for .NET 4.5, probably because it was only really done for Reporting Services and maybe because of test coverage. However, this code is used by the SQL Server provider, and is useful to other providers to the extent that I've heard they sometimes use Reflection to get at it. This change makes these remaining classes public so that they can be used by the SQL Server provider without using internals.

In keeping with other similar cleanups I also deleted some code that was #ifdefed out. If we need this code in the future it will still be in the history.