IfYou'reLazyYou'llAlwaysBeSingle (Replace uses of Singleton with thread-safe Lazy)
I spent some time trying to test this change but it became to complicated due to the static and singleton nature of the members involved. Therefore after talking to Diego we decided to just make this change based on inspection of the Singleton class, usage, and knowing how Lazy works.

One place (in shaper) I determined that the Lazy didn't have to be thread-safe. The rest need to be thread safe and so remain that way.