xUnitIndependenceDay - breaking the dependency between SampleProvider tests and xUnit package that lives in the EntityFramework branch.
SampleProvider is supposed to be a standalone project but it had a dependency the xUnit package that lives in the EntityFramework branch. To break this dependency I removed the HintPath for Unit.dll and enabled restoring NuGet packages for the solution.

Verification:
The solution builds and all tests pass after git clean -xdf