IndianaJonesAndTheRottenTreasuriesOfExtendedProvider - deleteing obsolete extended provider unit tests
Extended provider is shipped with some tests. These tests are using a mixture of MSTest and an
ancient mock library that is shipped in SDK. Due to these dependencies it is hard to make them
compile and after compiling they 2 (out of 3) won't pass due to setup issues. Given that there
are only 3 tests that are not really testing changes made for the EF sample provider I decided
to get rid of these tests. I already smoke tested the DDEX provider and will do some more manual
testing.