ProviderOfTheWeak (EF6 changes to test TryGetProviderInvariantName fix for .NET 4.5.)
(Note that I am making these changes in EF6 so that I could test them easily. I will then move the tested code into .NET 4.5 and do manual testing.
Also note that the code in EF6 will not stay as is--I will merge the two copies of this code and do additional cleanup. The changes here are done in such a way that they can be transfered easily to the .NET 4.5 code base.)

Bug:
Dev11 412714: EntityConnection.ConnectionString fails when a custom DBProviderFactory within an unsigned dll is used
http://vstfdevdiv:8080/web/wi.aspx?pcguid=22f9acc9-569a-41ff-b6ac-fac1b6370209&id=412714

The bug reported is that looking up a non-strongly named provider invariant name fails. The underlying issue is that the assembly matching code that Andrew made more robust in EntityFramework.dll never made it to the version of this code used in the core libraries. This code therefore also has the issue with generic providers, although this will show up in different situations because the code in the core libraris is called in different scenarios.

The fix is to use Andrew's code from EntityFramwork.dll. Most of the changes here are to test that that code is working for the weakly named assembly case.

CR: Andriy