WelcomeToProviderland (Allow EF6 provider to be specified in config file.)
This change adds the initial configuration support for EF6 providers. At this point it allows only specification of a provider type for a given invariant name for the entire app. The EF6 provider must follow the existing ADO.NET convention of including a static field or property called Instance that provides the Singleton instance of the provider.

Remaining work:
- Allow override for different context types
- Provider caching
- Integration of Spatial and Migrations services
- Write post for FWLink explaining how to get and use EF providers
- Allow setting the provider manifest token in the provider config