Merging from EF5:
ProppingUpTheScaffolding (Update DbContextInfo to allow configuration and provider info to be supplied at the same time.)

This allows scaffolding to create a Code First or Database First model without hitting the database. All the code to do this was already present but there was no DbContextInfo constructor that allowed both things to be specified. This code adds that constructor.