Making the materializer more testable.
Converted some static factory methods to instance methods on factory classes. This is needed to be able to override them with methods that provide mocks. Also it makes it easier to move to the dependency injection container.

Split Translater into CodeGenEmitter containing the static members related to codegen and TranslatorVisitor containing the visitor implementation. This is to separate responsabilities of the class and to hide the visitor implementation as it shouldn't be accessed in ways other than through the factory.

Repurpused the remainder of Translator to replace ShaperFactoryCreator. DRY: ShaperFactoryCreator and static factory methods on Translator that do the same thing.

Changed some uses of ReadOnlyCollection<KeyValuePair<ObjectParameter, QueryParameterExpression>> to IEnumerable<Tuple<ObjectParameter, QueryParameterExpression>> to better reflect the underlying semantics.

Added basic unit tests for the affected code.