(EjectingIntoSpace) Allowing dependency injection to be used to specify/override the provider spatial services
In EF5 it is not possible to use the DbGeometry and DbGeography factory methods for anything other than SQL Server or the limited default provider, and even this cannot be configured. This change makes spatial provider lookup first check the dependency resolver before falling back to ask the EF provider.

Most of the changes here are refactorings in the spatial code to make it possible to test that this actually works:
- SqlTypesAssembly was split to allow loading of the SQL types to be mocked
- SqlSpatialServices was changed to accept the loader
- DbProviderServices was changed to allow a resolver to be passed to it. By default it gets the resolver from the DbConfiguration in use. This allows the provider to resolve dependencies using EF configuration.