TypeTango (Remove GetTypesSpecial and update usages of GetTypes to GetAccessibleTypes)
When GetTypes is called on an assembly while running in partial trust it will fail (throw) if any of the types cannot be loaded. This means for cases like Code First discovery and O/C mapping you will not be able to use some assemblies in partial trust even if there are no problems with any of the types you actually want to use. We ran into this with the unit tests at one point.

Fortunately the exception thrown contains all the types that could be loaded, so by catching and returning the filtered list we are able to keep going. This is what the GetAccessibleTypes extension method does.

We had previously (June CTP timeframe) also added some code to special case getting types from the EF assembly. The underlying issuing causing this has now been fixed (see Dev11 205774) so this method is now no longer needed and has instead been replaced with GetAccessibleTypes.