AllYourTestBaseAreBelongToUs (Update tests to ensure TestBase static constructor is called when tests do some EF thing without a context)
We have a lot of tests that exercise EF without using a derived context type. If one of these tests runs and then a test that uses a derived context runs then the first test must have set the appropriate DbConfiguration. This is done by the static constructor on TestBase. This was failing in two ways that are now fixed:
1. Some tests that don't use a derived context were not using TestBase.
2. Some test classes use a static constructor which then gets executed before the TestBase static constructor.