InjectResolveGenerate (Allow Migrations SQL generation to be injected.)
This is the first part of the change to stop special-casing SQL Server and SQL Server Compact when doing database creation from DbContext. We were using Migrations SQL generation for these two but we couldn't do so for others because there was no way to register and find a SQL generator for a different provider. Now the SQL generator is resolved by DbConfiguration so that other generators can be injected. This will allow the Migrations SQL generation to be registered in the config file although that change has not been made yet.

Also included here is a refactoring of the root resolver to work as another Chain of Responsibility so that root resolution is not all grouped together into the single root resolver class.