MigrateMeMrNuGet (Allow Migrations SQL generator to be registered in the config file)
This change allows a provider writer to create a Migrations SQL generator and register it in the config file along with their EF provider.

No method has been added to DbProviderServices to get the generator and no property exposes the generator in DbConfiguration, but we could still choose to do those things later if desired.