TheDarkThreadRises (Make DbConfiguration and DbDependencyResolver uses thread-safe and add caching)
DbConfiguration and dependency resolvers can be accessed by multiple threads at the same time and hence need to be thread-safe. These changes do that using built-in data structures where possible and also some custom code. In addition some lookups (providers and app config lookups) are now cached.

While working on this I realized that setting a SQL generator in a Migrations configuration would set it for the app-domain, which is breaking change over what we were doing in EF5 so this has been changed back to being local with the app-domain (config file) setting only kicking in if the local is not set. I also realized that SQL generators are not thread safe so now a new instance is returned each time one is required which more closely matches how they were used before.