Query pipeline async implementation.
Added non-generic IDbAsyncEnumerator to be used in non-generic API instead of IDbAsyncEnumerator<object>. This is needed to avoid using reflection in the ForEachAsync implementation that was due to the fact that IDbAsyncEnumerator<T> can't be cast to IDbAsyncEnumerator<object> when T is a value type.
Changed Database.SqlQuery to return DbSqlQuery so that it could be enumerated asynchronously (without tracking). DbSet.SqlQuery will now return DbSqlSetQuery that is tracking by default.
Changed the DbSqlQuery.GetEnumerator method to execute the query lazily upon enumeration to match the async counterpart behavior. This is needed for IDbAsyncEnumerable implementation because GetAsyncEnumerator is a synchronous method that shouldn't perform IO operations.
Rearrenged MockHelpers so that they are contained in the same namespace as the corresponding type.
Unit tests will be added in a different commit.