ConfigTakesGold (Allow DbConfiguration type to be specified in the config file.)
This change allows the DbConfiguration type to use to be specified in the config file which in turn disables all discovery mechanisms and checking for configuration types in the same assembly.

Note that the the name of the element in the config file is "dbConfiguration" because it cannot be anything start with "config".