LetYourFeetBreathe (Stop throwing if DbConfiguration type is not found in same assembly as context type)
This change removes one of the foot protection checks that helps make sure that configuration is discoverable at design-time. We would previously throw if you explicitly set a DbConfiguration type to use in code (not from config) and that configuration was not in the same assembly as your context type. This helped prevent against causes where we would use the config at runtime but not be able to discover it at design-time.

The problem was that it causes exceptions to be thrown for contexts in assemblies completely outside of the app developer's control--for example, let's say that an MVC assembly had a context type in it--boom! There was no way to fix this, but even if we introduced a way to fix it it would be a pain and would force people to fix something that wasn't actually even a problem. So the conclusion is to be less protective in this case.