ManifestDestiny (Allow provider manifest token to be obtained without hitting the database)
The normal way to get the provider manifest token usually involves hitting the database to find out which version it is--for example, the SQL Server provider does this. This change introduces a service that can use any information in the connection to return a manifest token potentially without hitting the database.

This is useful both to avoid the database hit and to ensure that a model is built for a given
database version even if the system may sometimes connection to a more recent version.

An implementation of this has been added for our tests to always return "2008" without hitting the database.