StaticT1R2/3 (Add static sugar to DbConfiguration class)
This change splits DbConfiguration into two classes--DbConfiguration containing the now static public API and not static methods expected to be called from an overridden constructor, and an InternalConfiguration class which maintains the existing singleton instance and associated instance methods so that they can be tested. All the code could go into DbConfiuration, but when I did that it started to get really messy.

I have not added static methods for GetService because I'm not 100% convinced of their value and it would require adding methods for the "real" method plus each of the three extension methods. If there is still a desire to do this we should discuss in the EF design meeting.