BackToTheFutureUnicorns! (Fix issue with ThreadStatic in UtcNowGenerator)
ThreadStatic fields with initializers only get initialized correctly on the first thread. Recommended practice is to use ThreadLocal<T>.

Functionally, the change doesn't affect anything because DateTime is a value type and having it start of with a default value worked just as well as starting with UtcNow. But the code now follows best practice and behaves in the way it appears to behave.

work item:  394