FullyLoadedNuggets (Build NuGet package with both net40 and net45 assemblies)
This changes the NuGet project so that it copies .NET 4 and .NET 4.5 assemblies into the net40 and net45 folders respectively. The command line build is also changed so that when packaging it builds first for the given configuration (Release by default) and then again explicitly for ReleaseNet40. This ensures that with the default command line build on .NET 4.5 you will get both types of assembly built. If you build with VS then you’ll need to manually build both configurations if getting both assembly types into the NuGet package is important, otherwise the only assembly type that will be updated is the one you’re working on in VS. If you build on .NET 4 then you won’t get the net45 assemblies.

This change required that the PowerShell assembly be dependent on all other product assemblies, which also fixes a bug where the provider packages were not being rebuilt without InternalsVisibleTo for packaging.

Also, fix config manipulation to use same version for net40 and net45.