IKnowYouKnowSoJustTellMeAlready (Get DbProviderFactory from DbConnection on .NET 4.)
The DbProviderFactories.GetFactory(DbConnection) method was made public on .NET 4.5 so that it could be called by EF without needing InternalsVisibleTo. However, on .NET 4 there needs to be a way to get this information without using the internal method which cannot be called under partial trust. This change introduces some heuristics similar to those used for the June CTP that attempt to find the provider in the same assembly as the connection. If this doesn't work we then fall back to looking at all registered providers and finally at all registered providers and assignable connection types.

All of this is implemented as a service such that even if what we have doesn't work for some provider setup it should still be possible to inject a different service implementation to do the correct thing for the provider in question. (Worse case would be to load up System.Data.Entity.dll and use its access to the internal method to get the result.)

The logic to find the provider is essentially the same as that used to find the provider invariant name and so this has be factored into a common class, which also helps with testing.