ICanDoThatToo (Implement IObjectContextAdapter on ObjectContext )
This was requested in the CodePlex discussion (http://entityframework.codeplex.com/discussions/391595) as a way to provide a common interface for both DbContext and ObjectContext from which the ObjectContext can be obtained in a common way. Not super-high value, but very easy to do and doesn’t impact the DbContext API at all.

work item: 473