Fix for #488 - Many-to-many relationship left and right keys flipped after Entity Framework 5 upgrade
Problem: Many to many FK configuration was failing to correctly determine the correct association end targets associated with the "left" and "right" configured column names.
Solution: Annotate the association source end with the nav. property info and use this to determine on which ends the left and right keys should be configured.

- Also fixed LeftKey/RightKey doc comments so Rowan doesn't have to patch his CF book :-)