Fix for 481: NotMapped on base property throws if derived type discovered first
Issue 1: When the derived type is discovered before the base type and the base type has a NotMappedAttribute on a property an exception is thrown.

Solution 1: Added a check for NotMappedAttribute on the property in the base type before throwing the exception.

Issue 2: While investigating the above issue I found that when the derived type is discovered before the base type the overriden properties aren't lifted resulting in duplicate properties on the derived entity type.

Solution 2: Changed the condition used when lifting properties to the same condition used when mapping the derived type after the base type.

work item: 481

Fixs