SpacedIn (Allow global spatial provider to be registered in config file)
This change allows the global spatial provider (which is used when creating stand-alone DbGeography/DbGeometry types) to be registered in the app.config/web.config file. This would allow a third-party provider that supports spatial to change the global provider to something other than SQL Server without code needing to be added to the application.