HeAintLightHesMyExpressionTree (Remove use of lightweight code generation)
This change replaces the use of Ref.Emit and lightweight code gen that we use to create property getter and setter and constructor delegates with expression trees compiled to delegates. The new methods have been tested to work in partial trust even for non-public types/members, at least if there is no trust boundary between the application code and the EF assembly--in other words, when EF is not GAC'ed. Use of Ref.Emit for proxies is not removed.

This change also includes the removing the use of RuntimeMethodHandle and RuntimeTypeHandle, which since .NET 4 just wrap TypeInfo and MethodInfo anyway and so do not provide any benefit.