DontPretendYouDontKnowMe - Using enum with multiple assemblies and derived types throws MetadataException
Fix for workItem 532 - MustFix: Using enum with
multiple assemblies and derived types throws MetadataException

When loading enum types we would not add them to the dictionary of loaded
types. As a result we were not able to resolve an enum type if the type
using it lived in a different assembly. The fix is to add the type to
the dictionary as we already do for complex and entity types.

Work Item: 532