WhatWhatWhatDrop (Cleanup empty database created when Migrations pipeline from initializer fails)
Fix for CodePlex 640: http://entityframework.codeplex.com/workitem/640

The issue here is that creating the empty database and running the Migrations DDL pipeline to create the schema are two separate steps. This means that creating the empty database can succeed but then creating the tables, etc. can fail, leaving an empty database.

This is essentially the same issue that we fixed in .NET 4.5 for the ObjectContext DDL pipeline and the fix is basically the same. That is, if we create an empty database but then running the rest of the pipeline fails then we will go back and cleanup (i.e. delete) the empty database.

Work Item(s): 640