HelpImLockedOut (Allow interception and modification of config before it is locked)
This is the first part of a change to allow applications and tooling to look at and change DbConfiguration after it has been constructed and set up but before it is locked. This part allows for the interception to be done from code by registering an event handler; it is intended that it will also be possible to add an interceptor using the config file.

The event provides an IDbConfiguration instance, which is really a DbConfiguration presented with a slightly different view. Using this you can lookup which services have been registered using the DependencyResolver property and then use AddDependencyResoolver to a add new resolvers that override the current services. You can also add a resolver that will override settings in the config file.

Note that an additional change is required to the way the obsolete DefaultConnectionFactory property is handled such that it mirrors the connection factory obtained from DbConfiguration