FAC163 (Make DefaultConnectionFactory property work with config overrides)
The DefaultConnectionFactory is obsolete but it is inevitable that some apps will still be using it. For those that are using it, this change makes it possible for the DbConfiguration to wrap and override a connection factory that has been set in this way such that things like profilers will work with it.