InvariantProviderNameForMasses
StoreItemCollection reads provider invariant name but does not store and expose it.
As a result when trying to serialize StoreItemCollection we need to pass the value
that needs to be set in the StoreItemCollection even though it should not be necessary.
Adding StoreProviderInvariantName to expose the value read from the Provider attribute
on Schema element or passed to other constructors that don't read Ssdl.