WrappersDelight (Make it easy to wrap an existing service before configuration is locked)
This change makes it easy to add an OnLockingConfiguration handler that can obtain the service that would have been returned for a given type and key and instead wrap or replace this with a different service.

The main change to allow this is that the handler now provides a snapshot of the services registered at the time the handler is called which can then be saved and used later to return whatever would have been resolved before wrapping it. The wrapping resolver doesn't need to have any up-front information about which keys may be used.

Also included here are:
- Removal of the IDbConfiguration interface with the members moved directly into the event args class instead.
- Addition of a sugar method 'WrapService' that makes it easy to wrap any service when handling the event. This is not required--everything can be done with an appropriately written resolver.