AssociationSetAssociatedWithAssociationSetDeadEnd - adding properties to deal with different ordering of association and association set ends
In CodeFirst pipeline there is an assumptions that AssociationSet Ends
are in the same order as Association Ends. In general this does not have
to be true. Due to this assumption when dealing with arbitrary StoreItemCollections
SSDLSerializer would write an invalid SSDL. The fix is to add properties for
SourceEnd and TargetEnd so that they return ends based on the name and not on
the order and use those in the SSDL Serializer.