YouMayTouchTheArtifacts (Remove explicit file permissions checks from metadata cache/loader)
Metadata is cached in the app domain. This change removes explicit checks that the thread accessing the cached metadata has permission to access the file from which it was previously loaded. This check kind of made sense in a heterogeneous app domain, but really doesn't make sense with current .NET security and with the current plan for EF partial trust where the EF assembly is running without a trust barrier between it and the app assembly.