PapaDontPreachToTheMaterializerGirl (Remove explicit security checks from materializer)
This change removes the explicit demands, asserts, and other checks that the materializer was doing. The idea here is that if the both the app assembly and the EF assembly are running in partial trust (without a trust boundary between the assemblies) then the materializer will work because access to the EF internals through the equivalent of Reflection is allowed. This should also allow mapping to privates, which was previously not supported. If the app tries to map to a member that it cannot access, then EF also can't access it, and materialization will fail--as it would before. If both the app and EF are in full trust, then everything will work as before.

I will be adding specific partial trust testing after more of the issues have been fixed.