SayYesToUnicorns (Replace use of IsValidLanguageIndependentIdentifier)
We were using CodeGenerator.IsValidLanguageIndependentIdentifier as part of the check for valid EDM names. This method cannot be called in partial trust. However, we were only using this method if it had already matched our own regex first. In writing tests for this it became apparent that if the string passed our own regex then it would always pass IsValidLanguageIndependentIdentifier unless it contained a character of category 'Other, Format' (Cf). I changed our regex to not allow this and we therefore don't need the additional check at all anymore.