ITrustYouEvenIfYoureLazy (Update proxy generator to work in partial trust)
This change updates the proxy generation so the proxy assembly works in the same way as the EF assembly. Also included here is re-enabling of existing partial trust tests and removing of security attributes from all three product assemblies.

There are two limitations with proxy generation with partial trust:
- We cannot add an AssemblyResolve event for the assembly, meaning that binary deserialization of proxies will not work. However, binary deserialization appears to fail in partial trust anyway.
- If the type implements ISerializable, then generation of the proxy type will fail. Actually, if the type implements ISerializable then it can't be used in partial trust anyway, so we actually fail as soon as the type is accessed--usually when attempting to set values onto the context's DbSets.

I also discovered a bug in binary serialization of proxies unrelated to this work. The bug was that we assumed that _entityProxyFactory instance was always non-null because it is set in the contructor. But deserialization doesn't call the constructor, so it was null.