ResolutionWhichWontFailByFeb1 - enabling passing dependency resolver to EntityCommand
EF runtime uses configuration to resolve dependencies. Once configuration is set it is
sealed and is supposed to be immutable. This does not work for the designer where the user
can add/remove edmx files multiple times and therefore can change providers many times
during the session. To talk to providers we use EntityCommand. Allowing to pass own
resolver to EntityCommand enables sending queries to multiple providers during one session.