Longview (Consolidate item collection setup)
My previous change broke use of the MetadataWorkspace in some multi-threading scenarios with a cached workspace. This was due to not fully understanding the haphazard nature by which item collections are set on the workspace. This change takes advantage of Lazy to handle deferred initialization of item collections in a consistent and thread-safe manner. This is coupled with new constructors for MetadataWorkspace that take delegates to set up the lazy initialization. The previous method of registering item collections (RegisterItemCollection) is now obsoleted in favor of these new constructors.