TheTruthTheWholeTruthAndNothingButTheTruth
Properties on EdmProperty that correspond to facets used to return null values if the facet was constant. It was handy for CodeFirst
but did not work that well for the designer where we would have to dig into Facet to get the value. It also felt wrong that the
convenience methods returned a different value from the value that would have been returned by the underlying facet. The change
is to always return actual value (be it const or the one that was set) and add properties that can be used to tell if the facet is constant.