ElevenApplesUpOnTop (Prevent stack overflow from recursive DbConfiguration discovery)
EnsureLoadedForContext is called from various places as soon as a context type is known to ensure that the correct DbConfiguration is found. However, sometimes the derived DbConfiguration itself can make use of a context type (such as when the Migrations initializer is used) in which case this would cause recursive calls into EnsureLoadedForContext resulting in stack overflow.

The solution is to make EnsureLoadedForContext not actually create the DbConfiguration instance but rather just save the type. The instance is then only create the first time that the configuration is used.