AGrandDayOut (Use Code First type information for o-space loading)
This change extracts the o-space type loading functionality so that it can be used independently of where the CLR types actually come from. This allows Code First to use its knowledge about CLR types to setup the mapping without needing to do any assembly scanning or type matching. This provides a small perf benefit for creating Code First context instances.

This change also allows types in the same assembly with the same name but different namespaces to be used with Code First. Nested types are not yet supported but this change also paves the way for Code First with nested types.